Fixes #276: Allow passing a sitegroup to all IDE commands. (#291)
* Fixes #276: Allow passing a sitegroup to all IDE commands.

* Replace in all relevant commands.

* Fix test.

* Adding output.

* Fix test.
